requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
gitpython>=3.1.0
openai>=1.0.0
subprocess32>=3.5.4; python_version < '3.2' 